# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Knodex API
  description: |
    REST API for Knodex — a Kubernetes Resource Orchestrator (KRO) visualization
    and management platform.

    Knodex provides a web interface for browsing RGD catalogs, deploying instances,
    managing projects and RBAC, and monitoring OPA Gatekeeper compliance.

    ## Authentication

    Most endpoints require a JWT Bearer token obtained via:
    - **Local login**: `POST /api/v1/auth/local/login` with username/password
    - **OIDC login**: Browser flow via `GET /api/v1/auth/oidc/login`

    Include the token in the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```

    Tokens expire after 1 hour (HS256). There is no refresh mechanism — re-authenticate
    after expiry.

    ## Enterprise Features

    Endpoints tagged **enterprise** require an enterprise license. In OSS builds they
    return `402 Payment Required` or `404 Not Found`.

    ## Rate Limiting

    Authentication endpoints are rate-limited per IP. Protected endpoints are
    rate-limited per authenticated user (default 100 req/min).
  version: 1.0.0
  contact:
    name: Provops
    url: https://github.com/knodex/knodex
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8080
    description: Local development server (Tilt)
  - url: http://knodex.knodex.svc.cluster.local:8080
    description: In-cluster service
...
paths:
...
  /api/v1/rgds:
    get:
      summary: List catalog RGDs
      description: |
        Returns a paginated list of ResourceGraphDefinitions marked for catalog
        inclusion with the `knodex.io/catalog: "true"` annotation.
        Filtered by project visibility and organization.
      tags: [rgds]
      operationId: listRGDs
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NamespaceQuery'
        - name: category
          in: query
          description: Filter by category
          schema:
            type: string
        - name: tags
          in: query
          description: Filter by tags (comma-separated, AND logic)
          schema:
            type: string
        - name: search
          in: query
          description: Search in name and description (case-insensitive)
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: sortBy
          in: query
          description: Sort field
          schema:
            type: string
            enum: [name, namespace, createdAt, updatedAt, category]
            default: name
        - name: sortOrder
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: Paginated list of RGDs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRGDsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    post:
      summary: Create a ResourceGraphDefinition
      description: |
        Creates a `kro.run/v1alpha1 ResourceGraphDefinition` in the cluster from
        a YAML payload — the RGD Builder's "Use this spec" deploy hand-off
        (Story 50.2). The payload is strictly kind-locked: anything that is not
        a ResourceGraphDefinition is rejected with 400 (the endpoint is never a
        generic manifest-apply). The `knodex.io/catalog: "true"` annotation is
        ensured so the created RGD surfaces in the Catalog. Requires the
        `rgds/*` `create` permission (serveradmin-only by default).
      tags: [rgds]
      operationId: createRGD
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRGDRequest'
      responses:
        '201':
          description: ResourceGraphDefinition created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRGDResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: A ResourceGraphDefinition with this name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/instances:
    get:
      summary: List instances
      description: Returns a paginated list of deployed instances, filtered by project visibility.
      tags: [instances]
      operationId: listInstances
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NamespaceQuery'
        - name: rgdName
          in: query
          description: Filter by RGD name
          schema:
            type: string
        - name: health
          in: query
          description: Filter by health status
          schema:
            type: string
        - name: deploymentMode
          in: query
          description: Filter by deployment mode
          schema:
            type: string
            enum: [direct, gitops, hybrid]
        - name: projectId
          in: query
          description: Filter by project ID
          schema:
            type: string
        - name: search
          in: query
          description: Search by name
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: sortBy
          in: query
          schema:
            type: string
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: Paginated list of instances
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListInstancesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/projects:
    get:
      summary: List projects
      description: Returns projects visible to the current user based on RBAC policies.
      tags: [projects]
      operationId: listProjects
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
    post:
      summary: Create a project
      description: Creates a new Project CRD in the cluster. Requires `projects:create` permission.
      tags: [projects]
      operationId: createProject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Project already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from local login or OIDC flow (HS256, 1hr expiry)

  # -------------------------------------------------------------------------
  # Reusable parameters
  # -------------------------------------------------------------------------
  parameters:
    Page:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    NamespaceQuery:
      name: namespace
      in: query
      description: Kubernetes namespace filter
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: missing Authorization header
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FORBIDDEN
            message: permission denied
...
    ServiceUnavailable:
      description: Service unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # -------------------------------------------------------------------------
  # Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum:
            - NOT_FOUND
            - BAD_REQUEST
            - UNAUTHORIZED
            - FORBIDDEN
            - SERVICE_UNAVAILABLE
            - INTERNAL_ERROR
            - VALIDATION_FAILED
            - RATE_LIMIT_EXCEEDED
            - METHOD_NOT_ALLOWED
            - CONFLICT
            - DEPLOYMENT_MODE_NOT_ALLOWED
            - GITOPS_ERROR
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Additional error context (field-level errors, etc.)

    # --- Health ---
...
    RGDResponse:
      type: object
      properties:
        name:
          type: string
        namespace:
          type: string
        description:
          type: string
        version:
          type: string
        tags:
          type: array
          items:
            type: string
        category:
          type: string
        icon:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        instances:
          type: integer
        apiVersion:
          type: string
        kind:
          type: string
        allowedDeploymentModes:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ListRGDsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RGDResponse'
        totalCount:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    CreateRGDRequest:
      type: object
      required: [yaml]
      properties:
        name:
          type: string
          description: Overrides the YAML's metadata.name (DNS-1123 subdomain)
        namespace:
          type: string
          description: Overrides the YAML's metadata.namespace (DNS-1123 label, defaults to "default")
        yaml:
          type: string
          description: The full ResourceGraphDefinition manifest (kro.run/v1alpha1 only)
        runId:
          type: string
          description: Originating RGD Builder run id — stamps the run's actionTaken field (best-effort)
    CreateRGDResponse:
      type: object
      properties:
        name:
          type: string
          example: webapp-stack
        namespace:
          type: string
          example: default
        kind:
          type: string
          example: ResourceGraphDefinition
        apiVersion:
          type: string
          example: kro.run/v1alpha1
...
    ListInstancesResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/InstanceDetail'
        totalCount:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    InstanceDetail:
      type: object
      properties:
        name:
          type: string
        namespace:
          type: string
        kind:
          type: string
        apiVersion:
          type: string
        rgdName:
          type: string
        status:
          type: string
        health:
          type: string
        deploymentMode:
          type: string
        conditions:
          type: array
          items:
            type: object
        spec:
          type: object
        createdAt:
          type: string
          format: date-time
...
    ProjectListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProjectResponse'
        totalCount:
          type: integer
    ProjectResponse:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        destinations:
          type: array
          items:
            type: object
            properties:
              namespace:
                type: string
              server:
                type: string
        roles:
          type: array
          items:
            type: object
        resourceVersion:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        destinations:
          type: array
          items:
            type: object
            properties:
              namespace:
                type: string
              server:
                type: string
...
security:
  - bearerAuth: []